ciao-controller, ciao-launcher, payloads: Rename Migration flag
The identifier "migration" was used throughout the migration code to
indicate one of two things.  An instance is being restarted, rather
than started, or an instance is being stopped rather than deleted.
This was a little confusing.  This PR renames the various migration
identifiers to stop and restart as appropriate.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>