ciao-scheduler: Set ErrorStartFailure.Restart appropriately
Scheduler needs to set the ErrorStartFailure.Restart field to match the
Restart field of the Start command that generated the error.  If it does
not do this a failure to restart a node generated inside scheduler, for
example if all nodes are FULL, could lead to the restarted instance being
deleted, which is not what we want.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>