ciao-scheduler: propagate node connect/disconnect events
NodeConnected and NodeDisconnected are not flowing correctly today.
Initial cluster formation, or controller disconnect/connect, can leave
the controller for a time without knowledge of nodes.  The controller
eventually gets information from statistics frames.

This patch does a better job of synthesizing connect/disconnect events
to the controller and will replay a set of connect events to a
controller when it joins so it immediate has awareness of connected
nodes.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>