ciao-controller: abstract the listNodes() function
The current listNodes() function is coded in a way that tells a story
that it is return a list of information on compute nodes.  That's not
quite true.  But it would be useful if it were.  And then we'd want to
also be able to list other subsets of cluster nodes.

Toward that end listNodes() is abstracted into a listSubsetOfNodes(),
which takes an additional ssntp.Role argument.  This is wrapped by three
simple invocations of listNodes(), listNetworkNodes() and
listComputeNodes() which return all, network and compute nodes
respectively.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>