ciao-down: clearcontainers: use 'go get' to retrieve sources
You now need to have cc-oci-runtime in your GOPATH for it to build,
so use 'go get' to retrieve the sources rather than 'git clone', to
help ensure the source ends up in the right place.
Note we use '-d' and '/...' to prevent warnings during download
as the cc-oci-runtime does not have default Go code to build or
install.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>