templateutils: Return errors instead of panicking.
There were a number of error cases in templateutils that resulted in a
call to panic.  This was the appropriate behaviour when the code was
hosted inside ciao-cli.  It is not the correct behaviour for a generic
package.  Consequently, some of these panicking functions have been
modified to return an extra error return value and the panics have
been replaced by returns.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>