templateutils: Make table and sort functions work with pointers
The sort and table functions were not derferencing pointers.  This meant
that they would work on slices and arrays of structs but not pointers to
slices or arrays of structs.

The sort function has also been modified to sort and return a copy of the
slice passed to it rather than modifying the slice itself.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>