templateutils: Invalid usage should not make OutputToTemplate panic
A new error type has been introduced.  Errors of this type are generated
whenever a usage error is detected by one of the functions added by this
package.  They are propagated by panic.  This commit modifies
OutputToTemplate to trap any panics and check to see whether the object
associated with the panic is a UsageError.   If it is, an error is returned.
Otherwise the panic is propagated.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>