templateutils: Ensure that Opt functions are idempotent
There was a bug which could have lead to a corrupted help message if a
given option had been specified twice in the call to NewConfig.

This commit also changes the behaviour of NewConfig to enable no functions
if it is passed no arguments.  Users who wish to use a config object and
to enable all functions can use the OptAllFns option.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>