templateutils: Throw template.ExecError instead of UsageError
This commit modifies the functions provided by templateutils to panic with
a template.ExecError object rather than a UsageError.  This change ensures
that template.Execute will not panic when one of our custom function returns
an error.  Previously, we were throwing a UsageError which was explicitly
trapped by OutputToTemplate but not when users called CreateTemplate
followed by Execute on the resulting object.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>