ciao-cli: Fix node, tenant and volume list commands
These command were broken by the PR that moved the template code into
a separate package, templateutils.  The old implementation of createTemplate
returned nil, nil if the template script was an empty string and the commands
listed above relied on this behaviour.  The templateutils version of
CreateTemplate returns an error if the template string is "" and this is what
was causing those commands to fail.  This commit adds a check to each of these
commands to ensure that the template script is not empty before caling
templateutils.CreateTemplate.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>