kubicle: Add kubicle a tool to setup k8s clusters
kubicle is a small command line tool that makes it easy to set up k8s
clusters in ciao.  You simply need to provide a base image, which must
be some version of Ubuntu server, and kubicle will do the rest.  In short
it creates workloads for k8s master and worker nodes, creates instances from
these workloads and joins these instances together to form a k8s cluster.
An external-ip address can be asssociated with the master node to provide
external access to the k8s cluster.

The tool accepts a number of options, such as the number of worker nodes
to create, configuration information for those nodes and the user accounts
to create on those nodes.

kubicle also supports a delete command that removes all the ciao objects
that compose the k8s cluster in the correct order.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>