sqlite3db: In tests correctly generate memory database paths
The paths used for the the in memory databases used in unit tests are
supposed to be unique for each run however they were being generated
inappropriately. The use of string() on an integer converts that
integere into the UTF-8 character which was breaking the uniqueness
constraint as sqlite3 was interpreting some names as the same. Adjusting
to use fmt.Sprintf() gives the intented result.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>