bat: factor node getter to support compute and network
BAT's GetComputeNodes() was really getting all nodes (network and
compute), because that's what the API provided via the cli.  Now that
is fixed, I'd like to add test functionality around network nodes, but
to do so the getter needs to support both.  The old GetComputeNodes()
body becomes getNodes(), which is wrapped by calls from a new
GetComputeNodes() and a peer GetNetworkNodes().

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>