kubicle: Only allow one k8s cluster per tenant
This commit prevents kubicle from creating multiple k8s clusters
per tenant.  The reason we do this is that if there are multiple
k8s clusters per tenant, it's not possible to delete them cleanly.

Once we have a way to label ciao objects we'll be able to remove
these restrictions.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>