kubicle: Derive pool name from uuid of master workload
Kubicle was using a fixed name for the ip pool.  As ip pools are
cluster, rather than tenant relative, this meant that it was only
possible to have one k8s cluster per ciao cluster when using
external-ips.  This commit fixes the problem by deriving the name
for the external-ip pool from the uuid of the master workload.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>