ciao-controller: Avoid raciness when adding tenant for workload creation
This change modifies datastore's AddTenant() to return a newly defined
error if the tenant ID is already in use. This allows the code in
controller to avoid a race between using GetTenant to find out if there
is a tenant AddTenant to add a new one.

This change also ensures that the lock is taken early enough and that
the lock is released on error paths (previously the lock was not
releases on error paths.)

Signed-off-by: Rob Bradford <robert.bradford@intel.com>