kubicle: Fix worker options and clarify memory units
This commit fixes the options that control the resources allocated to the
worker VMs.  Previously, kubicle allowed users to specify the worker resources
on the command line but it ignored them, using the resources specified for
the master instead.  This has now been fixed.

The commit also clarifies the units used for memory and disk space of the
k8s VMs, both in the code and the help messages.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>