ciao-controller: delete CNCI if it is no longer needed
After deleting a tenant instance, wait 5 minutes and then see
if the CNCI is no longer needed. If it is not, delete it.
We wait 5 minutes to avoid the startup and tear down cost of
the CNCI if the user is going to immediately schedule more
work.

In order for us to be able to delete CNCIs, we need to start
storing the CNCI instance data into the datastore. However, we
do not want the CNCI to be visible to the tenant when querying
their instances or operating on them, so create a new call
to get Instances which specifically only gets Tenant visible
instances (i.e. not the CNCI). When getting instance info on
behalf of the API service, use the call to GetTenantInstance.
When getting instance info in other cases, use GetInstance,
which will also include CNCI instances.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>