ciao-down: Persist instance state properly
All instance state is now persisted properly in a yaml file.  This
means that when you restart an instance you get the same settings
that you specified when you first created the instance via prepare.
It is now possible to specify any number of directories to mount
in the VM and ports to map.  The attributes of the default paths
can be overriden on the command line.  For example,

--mount hostgo,none,$HOME/go

changes the security model of the GOPATH from passthrough to none.

Fixes #1261

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>