singlevm: Remove CIAO_CA_CERT_FILE environment variable
This environment variable is designed to hold the CA certificate which
the HTTP certificate used by ciao-controller is signed with. However
this was being set to the SSNTP CA certificate so was not having the
desired result.

Since the HTTPS certificate needed to talk to keystone is installed into
the system certificate store it is not necessary to set this variable.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>