ciao-down: Workloads now specify instance defaults
In addition to being able to specify the cloud-init file used to bootstrap
a new VM, the workload yaml files now contain two new sections.  The first
section contains invariants about the VM created from the workload such
as the image off which it is based and its hostname.  The second contains
dynamic instance data that can be altered on every boot, such as the number
of CPUs the VM uses and the ports mapped.  The final section contains the
cloud-init file.  If only one section is present it is assumed to the be
cloud-init file and default values are used for the other two sections.

Workload is now a multi-doc YAML file.  Sections should be separated from
each other using the normal '---' and '...' yaml delimeters.

Fixes #1106

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>