ciao-controller: Don't start a CNCI for "public" tenant
The tenant with ID "public" exists in the database to satisfy the
requirement that workloads are associated with a tenant. This
pseudo tenant is thus used to signify that this workload is public.

However as no instances are going to be created with this tenant it is
not necessary to launch a CNCI for the tenant.

Fixes: #1324

Signed-off-by: Rob Bradford <robert.bradford@intel.com>