ciao-controller: Use memory backend for transient database
For performance controller has separate databases for frequently updated
statistics that are not critical for operation and for state that needs
to be persisted long term. The former being stored in /tmp to avoid
filesystem performance issues. However /tmp is not guaranteed to be
tmpfs (and on Ubuntu 16.04 is not) instead this change switches
controller to use the memory backend to sqlite.

Fixes: #1330

Signed-off-by: Rob Bradford <robert.bradford@intel.com>