ciao-controller: Perform graceful shutdown on SIGINT/SIGTERM
This change modifies controller to shutdown gracefully upon reception of
the SIGTERM or SIGINT signals. It does this by modifying the HTTP
servers for API entry points to use the new Go 1.8 http.Server
interface. When each server is created it is added to a slice controlled
by controller. Upon reception of a signal the Shutdown() function on
each server is called resulting in the ListenAndServeTLS() functions
returning. When all active server goroutines are completed the Wait() on
WaitGroup returns and controller shutdown code can execute.

Fixes: #1326

Signed-off-by: Rob Bradford <robert.bradford@intel.com>