ciao-cli: Provide a hint to the user if they request a /32
Direct the user to user to use the address method for adding IP
addresses to the pool rather than specifying a /32 subnet for a single
address. The rationale for this is that subnets and IP addresses exist
as separate entities and cannot be mixed e.g. add as address, remove as
subnet.

Fixes: #1290

Signed-off-by: Rob Bradford <robert.bradford@intel.com>