kubicle: Install k8s version 1.6.7
There is a race condition in k8s 1.7.0 that prevents it from working with
kubicle.  Sometimes the worker nodes are unable to join the cluster.

The k8s bug is here:

https://github.com/kubernetes/kubeadm/issues/335#issuecomment-312962574

and will be fixed in k8s 1.7.1.  In the meantime we fix the k8s version
to 1.6.7 which is known to work well.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>