docs: Remove broken link from the README.md
The ciao README.md contains a link to the old very outdated cluster installation
instructions.  To make matters worse, the link doesn't even exist any more.
This commit simply removes the reference to the broken link.  We can add a new
link once we have a document describing how to set up a cluster using ciao-deploy.

I've also modified the SingleVM link so that it points directly to the main
SingleVM page.

Fixes #1335

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>