ciao-deploy: Introduce "ciao-deploy bat-create-workloads"
This change introduces a new command to create the workloads needed for
BAT in a configured cluster. It will download the files from the http
server if necessary, create the image in the cluster and then create
the desired workload.

Images are cached into $HOME/.cache/ciao/images

Extra workloads not strictly required for BAT are added with
--all-workloads.

This PR also introduces a new dependency of cobra, a command line
management tool.

The command is run with "ciao-deploy create-bat-workloads". Currently no
output provided except in the case of error.

It does not yet support injecting SSH keys or passwords into the
cloud-init for the created workloads. That feature will be added in a
later commit.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>