ciao-down: Rename instance to VMSpec
Just just workloadSpec, having the Spec suffix signals this structure is
serialized to the state file.

While instance is understandable, I'd like to introduce a "proper" VM
object in the future that is a 1:1 mapping with created VMs. For this
reason, anything associated with created VM should have VM in the name.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>