ciao-down: Fold the second document into the first one in the state file
At the moment some of our workload files have an empty 2nd document to
be used to save per-instance data. This is a bit awkward for people that
want to write workload files. Removing it will remove the need to
explain the little quirk.

This commit keeps compatibility with created VMs by being able to load
the previous version of state files. It will "migrate" them to the new
state structure when saving though.

Light testing indicates it seems to work as intended and we even have
some test coverage!

In the future, when we decide to break compatibility to support multiple
VMs at once, we can decide to drop the support for 1 and 3 documents in
the workload/state files and drop quite a bit of code.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>