ciao-deploy: Cleanup upon error for create-bat-workloads
If create-bat-workloads generates an error at any point, e.g. because a
long running task has been cancelled by user interuption try and cleanup any
changes that the tool has made to the system. It's important to note
that only images that the tool has downloaded on this run will be
deleted.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>