datastore: Remove asynchronous database updates
The datastore codebase updated the database asynchronously using go
routines. Unfortunately this would result in the error being lost as the
error from the database functions would be disregarded. This patch
removes the go routines for the database updates and for the errors
returned by the database either propagates it upstream if the calling
function returns an error or at the very least makes sure it is logged.

Further checks need to be made to ensure that the datastore cache is not
now out of sync with the database if it fails. Those issues will be
addressed in another patch.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>