ciao-deploy: Implement master node setup and update
This change introduces two commands to ciao-deploy. "ciao-deploy setup"
is used to configure the master node for participating a in a ciao
cluster. "ciao-deploy update" updates the software currently running on
the master node.

After provisioning the master node it is necessary to join compute and
network nodes to the cluste for it to be usable.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>