auth: Add basic authentication provider
This provides a mechanism to look up the username/password inside the
datastore and validate if the credentials are valid.

This is accessed by a http.Handler implementation that will attempt to
use the datastore to authenticate and if that fails will drop back to
using the next handler (e.g. osIdentity.Handler).

Signed-off-by: Rob Bradford <robert.bradford@intel.com>