ciao-controller: Respect http.Error()s requirements
http.Error() documentation says "Error replies to the request with the
specified error message and HTTP code. It does not otherwise end the
request; the caller should ensure no further writes are done to w. The
error message should be plain text."

This change ensures that in all codepaths where http.Error() is used
nothing else is written to the the http.ResponseWriter. In all cases
this is achieved by returning early.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>