ciao-controller: Use routes and handler from the real API for testing
When testing use the router and the HTTP handler that is used in the
live code when testing. This allows for more advanced testing by
supporting the handling of HTTP context variables for e.g. privilege and
by allowing the matching of url parts.

The bulk of this change is substituting valid UUIDs into the test data.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>