ciao-controller: add a CNCI controller to manage CNCI lifecycle
This commit adds a CNCI controller to help manage the CNCI lifecyle,
and allows for multiple CNCIs per tenant (1 per subnet).

Create and implement CNCI controller interface which supports starting
a CNCI per tenant subnet. When the subnet is no longer in use (the last
host IP is deleted from this subnet), delete the CNCI after a 5 minute
delay.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>