ciao-controller: CNCI launch failure should not crash
Failure to launch a CNCI was causing a crash.  Not when the failure
was detected but when an attempt was made to re-launch a new CNCI
for the same subnet.  This issue is fixed by this commit.  We simply
delete the information we have about a CNCI when we get a start failure.
This allows a new CNCI to be launched

Fixes https://github.com/01org/ciao/issues/1417

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>