ciao-cli, templateutils:  Switch ciao over to tfortools
The templateutils package has been moved to its own directory called tfortools.

github.com/intel/tfortools

This commit removes the old templateutils package from the top of the ciao
tree and vendors the new tfortools package.  ciao-cli is modified to use
tfortools.

The version of the code in the tfortools package is more up to date than what
we currently have in templateutils so switching to tfortools should fix some
issues with and enhance the ciao command line.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>