ciao-launcher: Implement evacuation
This commit partially implements the evacuation command in launcher.
When launcher receives an evacuation command it kills all of its
running instances and removes them from the overseer.  It is also
supposed to set the node into some sort of maintenance state but this
is not currently done.  We need to add the new maintenance state to
SSNTP first.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>