payloads, testutils: Add Restore command
This commit adds the Restore command to the payloads and testutil packages.
As the two packages depend on each other it was difficult to split up
these commits.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>

testutil: Add Restore command

Add test infrastructure for the Restore command.  Some refactoring of
SsntpTestServer.CommandNotify was needed to get the code to pass
gocyclo.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>