ciao-controller, ciao-cli: Reimplement evacuate and restore
Previously, evacuate and restore worked by sending either a DELETE or
a PUT command to the /node endpoint.  This didn't feel quite right
as we weren't really deleting the node.  We were just changing its
status.  This commit implements both commands using PUT on the
/node endpoint.  Both commands supply a json document containiing
the desired status of the node being operated on.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>