Implement support for modifying tenant attributes
Implement an api in controller to allow the admin to set 2 attributes
of the tenant; the name, and the number of bits in the subnet mask.
Allow a privileged user to list all the tenants that the controller
has saved in the datastore.

Allow a non-privileged user to obtain the attributes associated with
the tenant they are authenticating with.

Implement all support in the cli for listing the tenants, and
for getting/setting the tenant attributes.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>