networking/docker_plugin: Add a function for deleting the db
This commit adds a small function to the docker plugin that deletes
the boltdb database that it maintains.  It is intended to be called
by launcher when launcher performs a hard-reset.  Right now the
db is not getting deleted during hard-reset and this creates issues
on clusters deployed using ciao-deploy.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>