ciao-launcher: Reset wipes all docker networking artifacts
This commit modifies ciao-launcher's hard-reset function to ensure
that all docker networking artifacts are wiped.  There were two
problems with the old code.

1. The docker DB was not wiped.
2. The docker networks were not wiped if docker networking could
   not be initialised.  Since a corrupt DB could prevent this
   from happening, hard-reset did not always delete the docker
   networks ciao had created.

This wasn't a problem in singlevm as the cleanup.sh script
wiped the docker networking artifacts.  But it is a problem when
using ciao-deploy, so it needs to be fixed properly in launcher.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>