networking: Fix docker plugin DB
As far as I can tell the docker plugin DB was not actually working, in
that it was not possible to restore the networking state from the DB,
after launcher had restarted if data had been stored in the DB before
the restart.  The issue seems to have been that the names of the
buckets were mixed up.  Switching the names around fixes the issue.

There should be no more

Docker Init failed: Decode Error

errors in the logs.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>