ciao-launcher: Add --mgmt-net and --compute-net options
These options were removed when the configuration package was added.
The problem is that we can't always rely on the information sent in
the configuration package from scheduler.  For instance when performing
a hard-reset launcher doesn't connect to scheduler.  In order to be
able to cleanup docker containers cleanly it still needs to initialise
the network and to initialise the network it may need the information
about the mgmt and compute networks.  If the network cannot be
initialised launcher will still clean up but it will be really slow.
This is evident on singlevm if you call ./cleanup.sh after creating
a container.

This commit then adds these options back in and allows them to override
the settings received from scheduler, if any.  The intended user of these
new options is singlevm.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>