ciao-controller: rework AddTenant() to include a tenant config struct
When adding a new tenant to the datastore, pass in the tenant config
rather than generate one from inside the datastore code. This allows
us to use a user provided config in the future. Also, add code to ensure
that the newly created tenant does not already exist inside the datastore.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>