Docs: Remove out of date or duplicated documentation
This PR removes documentation that does really serve any real purpose,
is out of date or is duplicated elsewhere such as in godoc or on
ciao-project.github.io.

The content of the remaining README.md files will need to be either
migrated elsewhere, deleted or perhaps be given leave to remain if
there is no other suitable place for these files.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>