README: Update the top level README.md file
This commit updates the top level README.md file to point to the
new webpage rather than the old out of date documents in the source
tree that it used to link too.  Some of the architectural information
was removed from the README.md file as this is not really the
correct place for this information.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>