ciao-launcher: Retain network settings for hard-reset
Launcher now persists the network settings it receives from scheduler
in the configuration package.  This is done so that at a later stage
if hard-reset is run, networking can be initialised on the node without
requiring any information to be passed in to ciao-launcher from outside.
hard-reset can take a very long time if networking has not been
initialised.  This is usually only an issue when containers have been
spawned on a node.

Fixes https://github.com/ciao-project/ciao/issues/1490

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>