ciao-launcher: Fix live attach of volumes with qemu 2.9
Qemu 2.9 and greater place a restriction on the length of identifiers used
to identify block devices.  If the ID passed in by the QMP agent is greater
than 31 bytes, qemu silently truncates it.  This truncation is only performed
on the node-name property used in blockdev-add and not on the id property
used in device_add.  Consequently, it is impossible to use block device IDs
of greater than 31 bytes when hotplugging volumes as the node_name and id
properties do not match.

Fixes: https://github.com/ciao-project/ciao/issues/1500

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>