ciao-launcher: Cleanup if device_add fails
The launcher code that attached volumes to running VM instances did not correctly
clean up when an attempt to attach the guest portion of the volume, i.e., device_add,
failed.  This commit fixes the issue by attempting to delete the host portion of the
device if device_add fails.  It also adds some addition logs around the code that
unmaps the rbd device.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>