ciao-controller: Don't update volumes on stats reception
There was a race condition in controller that was leading to intermittent
storage BAT test failures.  The issue occured if a STATS update was received
for an instance while controller was trying to attach a volume to that
instance.  The STATS update does not contain information about the volume
if launcher has not yet processed the Attach command.  Controller was
noticing that the volume was missing from the STATS and mistakenly deleting it.
This is no longer necessary now that live volume detatch has been dropped as
a feature.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>