bat: Run some image BAT tests as non-admin user
Previously only the admin user could create images however that is no
longer the case and the BAT tests have not been updated to reflect that.

The change in the bat package to support this added a new boolean
parameter to all bat image functions to indicate whether to run as admin
or not. This necessitated changes to the bat tests but also to
ciao-deploy that uses the bat package to upload images.

Fixes: #1520

Signed-off-by: Rob Bradford <robert.bradford@intel.com>