ciao-cli: move core client code to new struct
Move the core client code that makes http requests and unmarshalls
the responses along with the certificate initialisation code into its
own struct and receivers.

This change also modifies the existing code to reference a new global of
this Client type (and fields within) rather than the multitude of
globals we had before.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>