ciao-launcher: Ensure VNICs get cleaned up after launch failure
There was some very dodgy code in launcher that prevented vnics from
getting cleaned up if a VM or a container launch failed.  The code was
also really convoluted.  It was using the instance state to determine
whether or not to delete a vnic but this was the wrong thing to do for
many reasons.  Pending instances can have vnics that need to be deleted
and also, by waiting until a failed instance is cleaned up to delete its
vnic we create a race condition with controller.  This commit fixes the
issue by cleaning up the vnic straight away if a launch fails and
by using a specially created instance variable that identifies whether
an instance's vnic needs to be deleted.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>