ciao-launcher: Add --osprepare and --roles options
This commit adds a new option to launcher, --osprepare.  When this option
is supplied launcher installs its dependencies and quits.  Launcher no
longer attempts to install any of its dependencies when this option is not
provided.  By default, launcher installs the dependencies it expects to find
on a compute node.  However, this behaviour can be overridden by the --roles
option.  --roles, is a string contaning a comma separated list of roles.
Two roles are currently defined, "agent" and "net-agent".

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>