ciao-scheduler: remove unused NetworkNode selection code
As a followup to 5ecb4a8c0a94b624220e1d4a5d9f23a876dbb6ee remove some
further unused code selecting a network node to run on based on the
network on the node.

If a CNCI needs to be associated with a specific node in the future it
can be done through the new node_id or hostname requirements.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>