ciao-deploy: Ensure tools' parent directories are present
This commit ensures that the directories into which newly installed tools
are to be copied exist by explicitly creating the directory.  Ideally,
this would be done with the install tool but as this tool is not yet used
in other parts of ciao-deploy, it's better to keep with the idioms
ciao-deploy currently uses.  A later PR will be need to modify ciao-deploy
to use install.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>