ciao-deploy: Ensure docker service is running on launcher nodes
This commit tries to ensure that the docker service is running on launcher
nodes by creating a dependency between the docker service and the launcher
service.  Starting the launcher service should now start the docker service
if it does not already exist.  Note launcher will still start if docker
is not present on the machine or cannot be started.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>