BAT: Fix a race condition in TestInstanceLimited
This test was creating and deleting instances but was not waiting for the
instances to be fully deleted before terminating.  This led to a race
condition in which instances could still be alive when the subsequent
test started.  As this test assumed no instances on start this race led
to an occasional failure of the following test (TestInstanceUsageAfterDenial).
This commit fixes the problem by ensuring that all instances have been
fully deleted before TestInstanceLimited terminates.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>